fixed bugs for middleware
