+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 .ds_store
+node_modules
 config
